removed secret key, good learning opportunity.
removing secret key . I don't think this is a big deal at this time as I am not intending on using this project with real data. 
learning curve, just found out about this secret key issue through a warning email from github. 
opportunity to learn it and not make the mistake again. will hide it in future projects, learning the process. Installed git-encrypt through homebrew for now.
As of today, I do not know if just removing the secret from this code here will have any impact on the repo etc. 
Will see.